backend/requirements.txt
bcrypt==3.2.0
Django==4.0
mysqlclient==2.1.0
PyJWT==2.3.0
simplejson==3.17.6
